Refactor: QrCodeManager 역할 분리 (#122)
Encoder, Decoder 생성
